Remove HOSTREGISTRY global object
Could have been causing sync issues on windows w/ subprocessing
